delete compet

+++ backend/inscriptions/src/index.js
@@ -46,8 +46,16 @@ app.post('/api/inscriptions', (req, res) => {
 });
 
 // Delete a database for a competition
-app.delete('/api/inscriptions/:competitionId', (req, res) => {
+app.delete('/api/inscriptions/:competitionId', async (req, res) => {
     const { competitionId } = req.params;
+    const check = req.query.adminId ? await axios.get(`${process.env.COMPETITIONS_URL}/api/competitions/${competitionId}/${req.query.adminId}`).catch((err) => { 
+        console.error(err);
+        return false; 
+    }) : false;
+    const admin = check ? check.data.data : false;
+    if (!admin) {
+        return res.status(403).json({ status: 'error', message: 'Forbidden' });
+    }
     deleteDatabase(`competition_${competitionId}`)
         .then(() => {
             res.status(200).json({ status: 'success', message: 'Competition deleted successfully' });
@@ -129,26 +137,18 @@ app.post('/api/inscriptions/:competitionId', async (req, res) => {
     const success_url = req.body.success_url || `http://localhost/competitions/${competitionId}?subPage=inscription&athleteId=${athleteId}&step=5`;
     const cancel_url = req.body.cancel_url || `http://localhost/competitions/${competitionId}?subPage=inscription&athleteId=${athleteId}&step=4`;
 
-    let admin = false;
-    const competInfo = (await axios.get(`${process.env.COMPETITIONS_URL}/api/competitions/${competitionId}`)).data.data;
-    if (competInfo.adminId == userId) {
-        admin = true;
+    const check = await axios.get(`${process.env.COMPETITIONS_URL}/api/competitions/${competitionId}/${userId}`).catch((err) => { 
+        console.error(err);
+        return false; 
+    });
+    const admin = check ? check.data.data : false;
+    if (!admin) {
+        return res.status(403).json({ status: 'error', message: 'Forbidden' });
     }
-    console.log(admin);
-
-    /*
-    try {
-        admin = req.query.admin ? (await axios.get(`${process.env.ADMINS_URL}/api/admins/${req.query.admin}`)).data.data : false;
-        // admin = req.query.admin ? (await axios.get(`http://admins-service:3000/api/admins/${req.query.admin}`)).data.data : false;
-    } catch (err) {
-        admin = false;
-    }
-    */
 
     if (admin) {
         userId = "admin";
     }
-
 
     // check body elements
     if (!userId || typeof userId !== 'string') {
@@ -262,13 +262,13 @@ app.put('/api/inscriptions/:competitionId/:athleteId', async (req, res) => {
     const success_url = req.body.success_url || `http://localhost/competitions/${competitionId}?subPage=inscription&athleteId=${athleteId}&step=5`;
     const cancel_url = req.body.cancel_url || `http://localhost/competitions/${competitionId}?subPage=inscription&athleteId=${athleteId}&step=4`;
 
-    let admin = false;
-   
-    try {
-        admin = req.query.admin ? (await axios.get(`${process.env.ADMINS_URL}/api/admins/${req.query.admin}`)).data.data : false;
-        // admin = req.query.admin ? (await axios.get(`http://admins-service:3000/api/admins/${req.query.admin}`)).data.data : false;
-    } catch (err) {
-        admin = false;
+    const check = await axios.get(`${process.env.COMPETITIONS_URL}/api/competitions/${competitionId}/${userId}`).catch((err) => { 
+        console.error(err);
+        return false; 
+    });
+    const admin = check ? check.data.data : false;
+    if (!admin) {
+        return res.status(403).json({ status: 'error', message: 'Forbidden' });
     }
 
     
@@ -425,9 +425,12 @@ app.put('/api/inscriptions/:competitionId/event/:oldEventName/:newEventName', as
     const { competitionId, oldEventName, newEventName} = req.params;
     const { adminId, isMulti } = req.body;
 
-    const check = await axios.get(`${process.env.COMPETITIONS_URL}/api/competitions/${competitionId}/${adminId}`);
+    const check = await axios.get(`${process.env.COMPETITIONS_URL}/api/competitions/${competitionId}/${adminId}`).catch((err) => {
+        console.error(err);
+        return false;
+    });
 
-    if (!check.data.data) {
+    if (!check || !check.data.data) {
         return res.status(403).json({ status: 'error', message: 'Forbidden' });
     }
 
@@ -460,13 +463,14 @@ app.put('/api/inscriptions/:competitionId/event/:oldEventName/:newEventName', as
 app.delete('/api/inscriptions/:competitionId/:athleteId', async (req, res) => {
     const { competitionId, athleteId } = req.params;
     let userId = req.body.userId;
-
-    let admin = false;
-
-    try {
-        admin = req.query.admin ? (await axios.get(`${process.env.COMPETITIONS_URL}/api/competitions/${competitionId}/admins?userId=${req.query.admin}`)).data.data : false;
-    } catch (err) {
-        admin = false;
+    console.log(athleteId);
+    const check = await axios.get(`${process.env.COMPETITIONS_URL}/api/competitions/${competitionId}/${userId}`).catch((err) => { 
+        console.error(err);
+        return false; 
+    });
+    const admin = check ? check.data.data : false;
+    if (!admin) {
+        return res.status(403).json({ status: 'error', message: 'Forbidden' });
     }
 
     if ((!userId || typeof userId !== 'string') && !admin) {
